api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
boto3==1.34.162
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0.post0
ulid-py==1.1.0

